requirements.txt
python-telegram-bot~=13.5
pynamodb~=5.0.3